refactor files, fix dropoff height, reduce bloom intensity, align text
