+++ src/themes/Alto/components/Dropoff.tsx
@@ -8,13 +8,10 @@ const Dropoff = () => {
   const { player } = useEnvironment();
 
   useFrame(() => {
-    if (camera.position.y < -50) {
-      player.controls.lock();
-      if (camera.position.y < -350) {
-        player.velocity.set(new Vector3(0, 0, 0));
-        player.position.set(new Vector3(0, 10, 32));
-        setTimeout(() => player.controls.unlock(), 2000);
-      }
+    if (camera.position.y < -15) {
+      player.velocity.set(new Vector3(0, -0.2, 0));
+      player.position.set(new Vector3(0, 1.5, 36));
+      setTimeout(() => player.controls.unlock(), 750);
     }
   });
